requirements.txt
torch
torchvision
torchio
pandas
numpy
neptune-client
nibabel
napari
matplotlib
seaborn
kornia
unet
tqdm
self_attention_cv